Data refactor part 1: Get rid of a lot of redundant classes, especially shift DataCollector's responsibilities to EventRepository
