+++ src/main/java/pl/edu/agh/arbeit/data/EventListener.java
@@ -1,6 +1,8 @@
 package pl.edu.agh.arbeit.data;
 
 import com.google.common.eventbus.Subscribe;
+import pl.edu.agh.arbeit.data.repository.DatabaseEventRepository;
+import pl.edu.agh.arbeit.data.repository.EventRepository;
 import pl.edu.agh.arbeit.tracker.Application;
 import pl.edu.agh.arbeit.tracker.trackers.ApplicationTracker;
 import pl.edu.agh.arbeit.tracker.trackers.SystemTracker;
@@ -12,7 +14,8 @@ import pl.edu.agh.arbeit.tracker.trackers.Tracker;
  **/
 public class EventListener {
 
-    private DataCollector dataCollector;
+
+    private EventRepository repository;
 
     public void subscribe(Tracker tracker){
         tracker.getBus().register(this);
@@ -23,18 +26,15 @@ public class EventListener {
     }
 
     public EventListener() {
-        dataCollector = new DataCollector();
+        repository = new DatabaseEventRepository();
     }
 
     @Subscribe
     public void event(Event event){
         System.out.println(event.getTopic() + " " + event.getDate() + " " + event.getType());
-        dataCollector.parseEvent(event);
+        repository.put(event);
     }
 
-    public DataCollector getDataCollector() {
-        return dataCollector;
-    }
 }
